Simplify protocol crate folder
