+++ bee-protocol/src/helper.rs
