--- bee-protocol/src/protocol/mod.rs
@@ -1,11 +1,0 @@
-// Copyright 2020 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-pub(crate) mod helper;
-
-mod metrics;
-mod protocol;
-
-pub(crate) use helper::Sender;
-pub use metrics::ProtocolMetrics;
-pub use protocol::{events, init, register, unregister};
