+++ bee-protocol/src/worker/heartbeater.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
+    helper,
     peer::PeerManager,
-    protocol::helper,
     tangle::MsTangle,
     worker::{PeerManagerWorker, TangleWorker},
 };
